initialcommit

+++ src/main/java/com/portal/repository/TuitionRequestRepository.java
@@ -9,7 +9,7 @@ import com.portal.entity.TuitionRequest;
 import com.portal.entity.Tutor;
 import com.portal.entity.User;
 
-public interface TuitionRequestRepository extends JpaRepository<TuitionRequest, Long> {
+public interface TuitionRequestRepository extends JpaRepository<TuitionRequest, Integer> {
     List<TuitionRequest> findByTutor(Tutor tutor);
     List<TuitionRequest> findByStudent(User student);
